face-align/requirements.txt
replicate
opencv-python
dlib
Pillow
scipy
matplotlib